Refactored room package
