+++ room/npc/nangarth.go
@@ -1,0 +1,8 @@
+package npc
+
+// NPCKris is a bro
+var NPCKris = NPC{
+	Name:   "Captain Kris",
+	Dialog: []string{"Ahoy matey!", "Have a good time in Alkos!", "I've had this here boat for 16 years"},
+	Speak:  SpeakDefault,
+}
